+++ src/restful/CourseSubjectRESTClient.java
@@ -3,18 +3,18 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package clients;
+package restful;
 
 import javax.ws.rs.ClientErrorException;
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.WebTarget;
 
 /**
- * Jersey REST client generated for REST resource:TeacherCourseSubjectFacadeREST
- * [entities.teachercoursesubject]<br>
+ * Jersey REST client generated for REST resource:CourseSubjectFacadeREST
+ * [entities.coursesubject]<br>
  * USAGE:
  * <pre>
- *        TeacherCourseSubjectRESTClient client = new TeacherCourseSubjectRESTClient();
+ *        NewJerseyClient client = new NewJerseyClient();
  *        Object response = client.XXX(...);
  *        // do whatever with response
  *        client.close();
@@ -22,15 +22,15 @@ import javax.ws.rs.client.WebTarget;
  *
  * @author z332h
  */
-public class TeacherCourseSubjectRESTClient {
+public class CourseSubjectRESTClient {
 
     private WebTarget webTarget;
     private Client client;
     private static final String BASE_URI = "http://localhost:32099/MazSolutionsServer/webresources";
 
-    public TeacherCourseSubjectRESTClient() {
+    public CourseSubjectRESTClient() {
         client = javax.ws.rs.client.ClientBuilder.newClient();
-        webTarget = client.target(BASE_URI).path("entities.teachercoursesubject");
+        webTarget = client.target(BASE_URI).path("entities.coursesubject");
     }
 
     public String countREST() throws ClientErrorException {
@@ -55,8 +55,12 @@ public class TeacherCourseSubjectRESTClient {
         return resource.request(javax.ws.rs.core.MediaType.APPLICATION_XML).get(responseType);
     }
 
-    public void create(Object requestEntity) throws ClientErrorException {
+    public void create_XML(Object requestEntity) throws ClientErrorException {
         webTarget.request(javax.ws.rs.core.MediaType.APPLICATION_XML).post(javax.ws.rs.client.Entity.entity(requestEntity, javax.ws.rs.core.MediaType.APPLICATION_XML));
+    }
+
+    public void create_JSON(Object requestEntity) throws ClientErrorException {
+        webTarget.request(javax.ws.rs.core.MediaType.APPLICATION_JSON).post(javax.ws.rs.client.Entity.entity(requestEntity, javax.ws.rs.core.MediaType.APPLICATION_JSON));
     }
 
     public <T> T findAll(Class<T> responseType) throws ClientErrorException {
