+++ src/restful/ExamRESTClient.java
@@ -3,18 +3,18 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package clients;
+package restful;
 
 import javax.ws.rs.ClientErrorException;
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.WebTarget;
 
 /**
- * Jersey REST client generated for REST resource:ExamSessionFacadeREST
- * [entities.examsession]<br>
+ * Jersey REST client generated for REST resource:ExamFacadeREST
+ * [entities.exam]<br>
  * USAGE:
  * <pre>
- *        ExamSessionRESTClient client = new ExamSessionRESTClient();
+ *        NewJerseyClient client = new NewJerseyClient();
  *        Object response = client.XXX(...);
  *        // do whatever with response
  *        client.close();
@@ -22,20 +22,20 @@ import javax.ws.rs.client.WebTarget;
  *
  * @author z332h
  */
-public class ExamSessionRESTClient {
+public class ExamRESTClient {
 
     private WebTarget webTarget;
     private Client client;
     private static final String BASE_URI = "http://localhost:32099/MazSolutionsServer/webresources";
 
-    public ExamSessionRESTClient() {
+    public ExamRESTClient() {
         client = javax.ws.rs.client.ClientBuilder.newClient();
-        webTarget = client.target(BASE_URI).path("entities.examsession");
+        webTarget = client.target(BASE_URI).path("entities.exam");
     }
 
-    public <T> T findExamSessionsBySubject(Class<T> responseType, String name) throws ClientErrorException {
+    public <T> T findExamsByStudent(Class<T> responseType, String name) throws ClientErrorException {
         WebTarget resource = webTarget;
-        resource = resource.path(java.text.MessageFormat.format("subject/{0}", new Object[]{name}));
+        resource = resource.path(java.text.MessageFormat.format("student/{0}", new Object[]{name}));
         return resource.request(javax.ws.rs.core.MediaType.APPLICATION_XML).get(responseType);
     }
 
@@ -61,14 +61,13 @@ public class ExamSessionRESTClient {
         return resource.request(javax.ws.rs.core.MediaType.APPLICATION_XML).get(responseType);
     }
 
-    public void create(Object requestEntity) throws ClientErrorException {
-        webTarget.request(javax.ws.rs.core.MediaType.APPLICATION_XML).post(javax.ws.rs.client.Entity.entity(requestEntity, javax.ws.rs.core.MediaType.APPLICATION_XML));
+    public <T> T findAllExam(Class<T> responseType) throws ClientErrorException {
+        WebTarget resource = webTarget;
+        return resource.request(javax.ws.rs.core.MediaType.APPLICATION_XML).get(responseType);
     }
 
-    public <T> T findExamSessionsByStudent(Class<T> responseType, String fullname) throws ClientErrorException {
-        WebTarget resource = webTarget;
-        resource = resource.path(java.text.MessageFormat.format("student/{0}", new Object[]{fullname}));
-        return resource.request(javax.ws.rs.core.MediaType.APPLICATION_XML).get(responseType);
+    public void create(Object requestEntity) throws ClientErrorException {
+        webTarget.request(javax.ws.rs.core.MediaType.APPLICATION_XML).post(javax.ws.rs.client.Entity.entity(requestEntity, javax.ws.rs.core.MediaType.APPLICATION_XML));
     }
 
     public <T> T findAll(Class<T> responseType) throws ClientErrorException {
@@ -80,8 +79,15 @@ public class ExamSessionRESTClient {
         webTarget.path(java.text.MessageFormat.format("{0}", new Object[]{id})).request().delete();
     }
 
-    public <T> T findAllExamSession(Class<T> responseType) throws ClientErrorException {
+    public <T> T findExamsBySubject(Class<T> responseType, String name) throws ClientErrorException {
         WebTarget resource = webTarget;
+        resource = resource.path(java.text.MessageFormat.format("subject/{0}", new Object[]{name}));
+        return resource.request(javax.ws.rs.core.MediaType.APPLICATION_XML).get(responseType);
+    }
+
+    public <T> T findExamByExamSession(Class<T> responseType, String id) throws ClientErrorException {
+        WebTarget resource = webTarget;
+        resource = resource.path(java.text.MessageFormat.format("examSession/{0}", new Object[]{id}));
         return resource.request(javax.ws.rs.core.MediaType.APPLICATION_XML).get(responseType);
     }
